Fix type consistency: uint for sprites and map ind
Change sprite parameters and map node indices from int to unsigned int
to accurately represent non-negative values. Eliminates unnecessary casts
and improves type safety.

- Sprite functions now take unsigned int sprite parameters
- All mn (map node) variables and related functions use unsigned int
- Coordinate variables in make_quick() use unsigned int
- DIST constant changed to 25U
- Removed unnecessary casts throughout

+++ src/amod/amod_structs.h
@@ -23,7 +23,7 @@
 #endif
 
 #define V_MAX         200
-#define DIST          25
+#define DIST          ((unsigned int)25)
 #define MAPDX         (DIST * 2 + 1)
 #define MAPDY         (DIST * 2 + 1)
 #define MAXMN         (MAPDX * MAPDY)
